requirements.txt
librosa>=0.10.0
numpy>=1.21.0
pydub>=0.25.1
lxml>=4.9.0
JPype1>=1.4.0
